renamed gs108e lib to netgear_plus
